CLOUDIFY-2023 Force retry of stateless pu instance if after 5 minutes the gsm did not retry to deploy it
svn path=/xap/trunk/openspaces/; revision=155485


Former-commit-id: ee16071492446ea75113aa0e2a23517977001490

+++ src/main/java/org/openspaces/grid/gsm/rebalancing/RebalancingUtils.java
@@ -943,7 +943,7 @@ public class RebalancingUtils {
     }    
 
     public static FutureStatefulProcessingUnitInstance restartProcessingUnitInstanceAsync(
-            ProcessingUnitInstance candidateInstance, Log logger, int timeout, TimeUnit timeUnit) {
+            ProcessingUnitInstance candidateInstance, Log logger, long timeout, TimeUnit timeUnit) {
         
         return relocateProcessingUnitInstanceAsync(
                 candidateInstance.getGridServiceContainer(),
